cleaned code
git-svn-id: https://vcs.maemo.org/svn/omweather/trunk@4540 5aa3d57c-301e-0410-a7f2-8fec96c11689

+++ qtomweather/qt-qml/configqml.cpp
@@ -93,7 +93,7 @@ ConfigQml::filename()
 }
 void ConfigQml::changestation()
 {
-    if (this->current_station_id() + 1 < this->stationsList().size()){
+    if ((uint)(this->current_station_id() + 1) < this->stationsList().size()){
        this->current_station_id(this->current_station_id() + 1);
 
    }else {
@@ -113,7 +113,7 @@ ConfigQml::refreshconfig(){
 void
 ConfigQml::updatestations()
 {
-    int i;
+    uint i;
     int success = 0;
     Core::Station* station;
